[PM-1594] Refactor Dialog and LunaModal
This change will move every Dialog UI related action to the Dialog
component:
 - If there is an async action on a dialog button, the dialog will
take show the respective button as 'loading' and disables the
other button.
 - If the action performed by clicking a dialog button throws
an error, the error will be shown in the dialog's footer.

LunaModals will be lockable. This will be achieved by using a new
state container called ModalLocker. A locked modal is not closable.
Also, the LunaModal will have 'destroyOnClose' attribute on by default.

Helper method 'wrapWithModal' is introduced to reduce boilerplate.
It wraps a component into a modal and passes its props to the wrapped
component.

Every affected component is refactored and simplified.