[PM-1759] [PM-1776] Add loading animation and fix responsiveness
Add loading animation for splash screen and fix its responsivness.

The animation is generating random images, so the splash screen
will be ignored from storyshot tests.

Create fullscreen decorator for storybook, which removes the padding
of the container elements so the story is shown in full screen. This
was needed for Splash Screen, which expects to be displayed in full
screen.

Cleanup unused images from code-base.

Other fixes:
 - Add id="main" to storybook's main element. Copying the App's main
element structure.
 - Extract beforeScreenshot delay time to constant in storyshots
 - Restrict the window's minimum height to 700px. This is the height
where the collapsed sidebar still show all its items.
 - Since the non-collapsed sidebar breaks under 800px, the sidebar
will collapse when the window's height is under 800px.