[PM-1963] Fix child process closing on Windows (#143)
* [PM-1963] Fix child process closing on Windows

Several problems were addressed in this change:

 1. There is an automatic restart of clients if some of them
is not running. Sometimes, this was triggered when Luna was closing.
This fix adds a flag, when Luna is shutting down, which is checked
before restarting the clients.

 2. When starting the backend on Windows, Luna creates 3 levels of
processes for both node and wallet. However, it only keeps track of
the highest one for each. When closing Luna, this process is killed,
but this action isn't propagated to its children. This change will
add a check to get the PID of the child process' child which runs
'java.exe'. This PID will be stored with the child process. When
Luna closes this child process it will kill the Java process too.

 3. Add wait before restarting clients.