[Dependencies] Upgrade antd
Following package changes were made:
 - Upgrade antd to latest (4.2.5) version
 - Icons in antd are now in separate package "@ant-design/icons"
 - @types/react and @types/react-dom were updated
 - @storybook/addon-info was never used, it is removed
 - tests were failing after the update, "jest-environment-jsdom-sixteen"
needed to be added and tests are now run with a special flag

Ant v3 to v4 global changes (see https://ant.design/docs/react/migration-v4 ):
 - Icons are now loaded per type
 - The default line-height changed, Luna will enforce
old line-height
 - Popovers now expect a single child only (wrapping every content
into a span if needed)

Luna specific changes:
 - BorderlessInput was a wrapper around ant's Input component.
Since ant's Input component's HTML structure completely changes
when using prefix or suffix for it, BorderlessInput will disallow
to use them. This includes the Password components visibilityToggle
which will be set to false always. Prefix on Input was used
in Burn activity, this is fixed.
 - DialogSeedPhrase was slighty changed to reflect ant's Autocomplete
component's HTML structure
 - WalletRestore test was changed to not use ant's classnames
instead to use Luna's class names when finding Seed input field.
 - SquaredSwitch was slighty changed to reflect ant's Switch
component's HTML structure
 - .ant-popover-arrow class override its color. Unfortunately,
with css `!important`, since what ant does is madness (the arrow's
design is defined for each of the 8 positions the popover can be).

There was some slight change in the buttons, which sometimes is
picked up by storyshots. It is hardly visible for the bare eye,
so it was kept as it is for now.