[PM-1697] Change DialogApproval to stop from continuing form
This commit wraps DialogApproval into a Form.Item. From now on,
every DialogApproval which has a set id property has to be checked
to validate the form. If it is not checked it will be shown
with red borders after validation. If a form validation fails
for a Dialog, it will show a general error message, which will
disappear if any field is updated.

The following components and related tests were affected and changed
to the new way of validation: ClaimWithKey, ClaimWithMessage, Exchange,
BurnCoinsGenerateAddress, WalletCreateDisplayRecoveryStep,
WalletCreateVerifyRecoveryStep, LogOutModal, SendTransaction (partially)