[Hotfix] Set main element a position: relative attribute
When positioning element with absolute in Luna's content, Sidebar
should be ignored. The easiest way to achieve this is to set the
main element (the container of all content) to position: relative.
Everything positioned absolutely will fallback at most to this
element.

This caused problems in GD when there was 'No wallet' and in PoB
for the testnet warning.