[Hoftix] Fix full amount button for redeem modal
DialogInput didn't trigger onChange when the value was changed
with the fill button. Removing fill button and moving logic
to redeem modal.

Also, full amount should consider the fee too. There will be
a useFullAmount flag which will make sure the actual full amount is
computed by substracting the fee from the whole balance. Since,
the fee estimation is dependent on the amount, it will be disabled
when in useFullAmount mode.