[PM-1938] Add proper logging for Luna
Add proper logging using electron-log. mainLog will be used for
the main process and renderLog for the renderer process. The log
file is formatted the way every entry is a JSON with timestamp,
level, and data.

console will no longer be permitted in Luna (either mainLog or rendererLog
should be used). This will be enforced by the eslint rule. It is set
to warning (so in dev mode we can still use console.log-s). However,
the pre-push hook will not allow warnings, so it will not allow to
push code with console.