[PM-2184] Showing DAG loading percentage
For some unknown reason, if the DAG's status message is not set
in the beginning, it's invisible for the renderer. Let's set it to
an empty string by default.

In `src/main/main.ts` when the `remote-get-global` event is processed
the return value is set to luna status with the correct DAG message.
However, in `src/config/renderer.ts` the `loadLunaStatus` method
returned luna status without the DAG message.