[Fix] Show in progress icon when hiding burns
Previously, when hiding burns, the action was instant by updating
the burn statuses correctly. However, the burn statuses are also
updated periodically too. This could lead to erroneous UX, where
a burn was hidden by the user, it was 'unhidden' again by a lagging
periodical update and hidden again with the next one.

From now on, burn status hiding will be left to the periodicall
update and Luna will show a loading icon in the meantime.