[Refactor] Introduce a global simplified Chain interface
Use a unified Chain interface across Luna for DST, BTC, ETH and ETC.
This new interface is a simplified version of the DisplayChain
interface which still had PoB speficic fields. The old Chain interface
(and all its usage) is now prefixed with 'Pob' prefix.

The clippedLogo property of the DisplayChain interface was removed
and replaced with logo (which now stores the clipped logo). This
affected the Token component's design.

DialogAddress component expected a full chain object, but now
it expects only the svgLogo property to know what icon to use.