[Fix] Downgrade react-scripts
react-scripts was previously upgraded to the latest version, so
the latest version of eslint-plugin-jsx-a11y could be used.

The latest version of react-scripts uses a buggy version of
eslint-plugin-imports which fails on Windows (see https://github.com/benmosher/eslint-plugin-import/issues/1643)

Since react-scripts needed to be downgraded, the following
packages also needed to be downgraded: babel-loader and eslint-plugin-jsx-a11y.