[PM-2294] Add ports to config, check if backends are already running
Check if backends are already running by checking if the ports are
already in use. Don't start Luna, when selected ports are in use.

Related fix:
Error box was fixed to fallback to electron's error box when the app
is not yet ready. In this case, electron cannot create a new dialog
and it fails. However, electron's original error box will fallback
to output its content to the console.