[PM-2400] Add ERC20 token support
Add ERC20 support

 - Add new menu item Tokens
 - In Tokens overview the user sees his tracked tokens or can add
new one clicking on 'Add New Token'
 - Adding a new token is done by providing the address of the smart
contract. The smart contract is validated if it is truly a ERC20
contract, this is done by checking if 'balanceOf' and 'totalSupply'
methods are present on the contract. The name, symbol, decimals
properties are prefilled from the contract if possible. If not
the user must provide them.
 - Tokens can be hidden and removed from the list.
 - Sending tokens is done per (transparent) account and the user
can choose which account he wants to use for the transfer. Only
accounts with available tokens are listed.
 - For receiving tokens any of the existing transparent addresses
can be used. For better navigation, the current balance is visible
in the transparent address list.

Related changes:

 - ERC20 contract abi was replaced with universal one which also
includes the name, symbol, decimals properties.
 - ShortNumber component and areFundsEnough utility method were
extended to support arbitrary decimal places.
 - Extracting a ScrollableModalFooter from ReceiveTransaction,
so it can be reused
 - Refactor ERC20 token tracking in wallet-state, making it generic
and extendable (some little changes were needed in PoB too)
 - Call Transaction Status return data will be only parsed (and
displayed in tx history) if the transaction failed. Otherwise,
it is not guaranteed that the data is a string.
 - Dialog component was slightly changed to prevent ant form's
automatical field validation on submit.

Other small changes:

 - Provide className for CopyableLongText component
 - Provide placement property for InfoIcon component
 - Move number type input design from DialogFee to global styles
 - Make DialogInput's label a ReactNode, and make possible
providing a Trans element for it
 - DialogMessage component was used as a whole box,
moving label out of the box
 - 'No Burn Activity' message in Burn Centre missed styling
 - GD and PoB layouts used the same margin, extract it to
a constant (and use it for Tokens)